fetch botometer data
