+++ package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "@faker-js/faker": "^7.6.0",
+    "botometer": "^1.0.12",
     "concurrently": "^7.6.0",
     "nps": "^5.10.0",
     "twitter-api-sdk": "^1.2.1",
