+++ test/botometer.test.ts
@@ -1,0 +1,11 @@
+import { botometer } from '../src/utils'
+
+describe('botometer', () => {
+  it('fetches botometer scores for a list of usernames', async () => {
+    const usernames = ['@r1oga', '@VitalikButerin']
+    const scores = await botometer.getScores(usernames)
+
+    expect(scores[0]).toMatchSnapshot()
+    expect(scores[1]).toMatchSnapshot()
+  })
+})
