+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src", "test"],
+  "include": ["src", "test", "types"],
   "compilerOptions": {
     "noEmit": true,
     "baseUrl": ".",
